Organize the code in two folder and added two more code file
