+++ C language Code/Adding two polynomials.c
