+++ C language Code/DEQueue-using-array.c
