+++ C language Code/active selection.c
