+++ C language Code/longest_sequence_consecutive_zeroes.c
