+++ C language Code/negative positive.c
