+++ C language Code/palindrome.c
