+++ C language Code/permutations.c
