+++ C language Code/quick.c
