+++ C language Code/searching_element_in_array.c
