+++ Cpp codes/Flatten binary tree to linked list.cpp
