+++ Cpp codes/count_of_keys.cpp
