+++ Cpp codes/merge_2array.cpp
